refactor: AnnouncementImgModal을 별도의 파일로 분리
