+++ service-manager/src/components/announcement/AnnouncementImg.tsx
@@ -1,6 +1,6 @@
 import { Dispatch, ReactNode, SetStateAction, useState } from 'react';
 import Close from '../../assets/close.png';
-import { Modal } from '@quokka/design-system';
+import { AnnouncementImgModal } from './AnnouncementImgModal';
 
 interface AnnouncementImgListProps {
   images: string[];
@@ -11,11 +11,6 @@ interface AnnouncementImgListProps {
 interface AnnouncementImgProps
   extends Pick<AnnouncementImgListProps, 'setImages' | 'isEditPage'> {
   image: string;
-}
-interface AnnouncementImgModalProps {
-  image: string;
-  isOpen: boolean;
-  setClose: () => void;
 }
 
 export function AnnouncementImgList({
@@ -83,39 +78,5 @@ export function AnnouncementImg({
         />
       )}
     </>
-  );
-}
-
-export function AnnouncementImgModal({
-  image,
-  isOpen,
-  setClose,
-}: AnnouncementImgModalProps) {
-  return (
-    <Modal
-      className="announcementImg-modal"
-      isOpen={isOpen}
-      onRequestClose={setClose}
-      overLayCss={{
-        display: 'flex',
-        justifyContent: 'center',
-        alignItems: 'center',
-        zIndex: 20,
-        width: '100%',
-      }}
-      contentCss={{
-        borderRadius: '1rem',
-        width: '800px',
-        position: 'relative',
-      }}
-    >
-      <button
-        onClick={setClose}
-        className="text-6xl absolute -right-16 -top-16"
-      >
-        X
-      </button>
-      <img alt="공지사항" src={image} className="mx-auto w-fit max-h-screen" />
-    </Modal>
   );
 }
